--- src/main/java/school/faang/amazonwarehouse/Category.java
@@ -1,8 +1,0 @@
-package school.faang.amazonwarehouse;
-
-public enum Category {
-    FOOD,
-    ELECTRONICS,
-    CLOTHING,
-    OTHER
-}
