--- src/main/java/school/faang/arrayrevert/Example.java
@@ -1,12 +1,0 @@
-package school.faang.arrayrevert;
-
-public class Example {
-    public static void reverse(int[] numbers) {
-        int temp;
-        for (int i = 0; i < numbers.length / 2; i++) {
-            temp = numbers[i];
-            numbers[i] = numbers[numbers.length - 1 - i];
-            numbers[numbers.length - 1 - i] = temp;
-        }
-    }
-}
