--- src/main/java/school/faang/cachingcaching/FrequentUpdateWeatherCache.java
@@ -1,12 +1,0 @@
-package school.faang.cachingcaching;
-
-public class FrequentUpdateWeatherCache extends WeatherCacheTemplate {
-    public FrequentUpdateWeatherCache(WeatherProvider provider) {
-        super(provider);
-    }
-
-    @Override
-    public boolean isCacheExpired(WeatherData data, long maxCacheAgeMillis) {
-        return true;
-    }
-}
