--- src/main/java/school/faang/cachingcaching/WeatherData.java
@@ -1,13 +1,0 @@
-package school.faang.cachingcaching;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class WeatherData {
-    private String city;
-    private double temperature;
-    private double humidity;
-    private long timestamp;
-}
