--- src/main/java/school/faang/cachingcaching/WeatherProvider.java
@@ -1,5 +1,0 @@
-package school.faang.cachingcaching;
-
-public interface WeatherProvider {
-    WeatherData fetchWeatherData(String city);
-}
