+++ src/main/java/school/faang/sprintone/amazonwarehouse/Category.java
@@ -1,0 +1,8 @@
+package school.faang.sprintone.amazonwarehouse;
+
+public enum Category {
+    FOOD,
+    ELECTRONICS,
+    CLOTHING,
+    OTHER
+}
