+++ src/main/java/school/faang/sprintone/arrayrevert/Main.java
@@ -1,0 +1,12 @@
+package school.faang.sprintone.arrayrevert;
+
+import java.util.Arrays;
+
+public class Main {
+    public static void main(String[] args) {
+        int[] arr = {1, 2, 3, 4, 5};
+        System.out.printf("Массив пришел такой: %s%n", Arrays.toString(arr));
+        Example.reverse(arr);
+        System.out.printf("Массив вышел такой: %s%n", Arrays.toString(arr));
+    }
+}
