+++ src/main/java/school/faang/sprintone/cachingcaching/FrequentUpdateWeatherCache.java
@@ -1,0 +1,12 @@
+package school.faang.sprintone.cachingcaching;
+
+public class FrequentUpdateWeatherCache extends WeatherCacheTemplate {
+    public FrequentUpdateWeatherCache(WeatherProvider provider) {
+        super(provider);
+    }
+
+    @Override
+    public boolean isCacheExpired(WeatherData data, long maxCacheAgeMillis) {
+        return true;
+    }
+}
