+++ src/main/java/school/faang/sprintone/cachingcaching/WeatherProvider.java
@@ -1,0 +1,5 @@
+package school.faang.sprintone.cachingcaching;
+
+public interface WeatherProvider {
+    WeatherData fetchWeatherData(String city);
+}
