+++ src/main/java/school/faang/sprintone/cachingcaching/WeatherService.java
@@ -1,0 +1,16 @@
+package school.faang.sprintone.cachingcaching;
+
+import java.util.Random;
+
+public class WeatherService implements WeatherProvider {
+    private final Random random = new Random();
+
+    @Override
+    public WeatherData fetchWeatherData(String city) {
+        double temperature = random.nextDouble() * 50;
+        double humidity = random.nextDouble() * 100;
+        long timestamp = System.currentTimeMillis();
+        return new WeatherData(city, temperature, humidity, timestamp);
+    }
+}
+
