+++ src/main/java/school/faang/sprintone/westeroslibrary/Book.java
@@ -1,0 +1,10 @@
+package school.faang.sprintone.westeroslibrary;
+
+import lombok.Data;
+
+@Data
+public class Book {
+    private final String title;
+    private final String author;
+    private final int year;
+}
