+++ src/main/java/school/faang/sprinttwo/ourbelovedhogwarts/SpellAction.java
@@ -1,0 +1,6 @@
+package school.faang.sprinttwo.ourbelovedhogwarts;
+
+@FunctionalInterface
+public interface SpellAction {
+    String castSpell(String spellName);
+}
