+++ src/main/java/school/faang/sprinttwo/ourbelovedhogwarts/SpellCaster.java
@@ -1,0 +1,8 @@
+package school.faang.sprinttwo.ourbelovedhogwarts;
+
+public class SpellCaster {
+    public void cast(String spellName, SpellAction action) {
+        String result = action.castSpell(spellName);
+        System.out.println("Результат заклинания: " + result);
+    }
+}
