--- src/main/java/school/faang/westeroslibrary/Book.java
@@ -1,10 +1,0 @@
-package school.faang.westeroslibrary;
-
-import lombok.Data;
-
-@Data
-public class Book {
-    private final String title;
-    private final String author;
-    private final int year;
-}
